Initial commit
Adds:

- basic web application listening on port 3000 using the [Negroni] web
  framework

- `/debug/vars` endpoint, exposed by [expvar], to see how the
  application was started (`cmdline`) and its memory usage (`memstates`)
  in JSON format

- an acceptance test suite using [Agouti] to drive Chrome browser using
  [PhantomJS]; for now it only checks that [expvar] is correctly
  exposing `/debug/vars`

Note that I'm using `http.DefaultServeMux` with Negroni (rather than
call `http.NewServeMux()` as the per the examples in Negroni's `README`)
since `DefaultServeMux` is the default used by `net/http` so there's no
point in creating a new mux. We will probably need to replace the
multiplexer from `net/http` with a more powerful routing library later.

The application is started inside the public `Run()` function so that it
can be called by the acceptance tests.

I chose the PhantomJS driver over Selenium and ChromeDriver for Agouti
because it was the fastest (approx. 1 second versus 4 and 2 seconds
respecitvely to start the driver, run one test and stop the driver).