Remove Run() function; use main() instead
The code to start a HTTP server was contained in the exported `Run()`
function so that it could be called before the tests start to run
(otherwise, there'd be no HTTP server to test).

This felt like a kludge, so move the code back to the `main()` function
and make the tests belong to the main package so that we can invoke
`main()` directly when setting up the test suite.