Use Negroni Logrus middleware first
Move the line that configures Negroni to use the Logrus logger so that
the logging middleware is the first to be called by Negroni.

This will ensure that the latency timings recorded by the Logrus
middleware are correct; the timer will start as soon as the Logrus
middleware's `ServeHttp()` method is run.