Copy negronilogrus package so we can pass a logger
Copy the `negronilogrus` package from
`github.com/meatballhat/negroni-logrus` to `middleware/negroni_logrus`
and modify it so that we can pass in an instance of `*logrus.Logger`.

This gives us more control over the Logrus logger used, and makes it
eaiser to pass that logger to packages without having to instantiate
`negronilogrus` first.

Also, modify `negronilogrus` such that:

- the package is now named `negroniLogrus`

- it no longer registers log entries with the `web` prefix (prefix is
  now empty)

- it no longer uses two fields to log latency; the `measure#web.latency`
  has been removed

- the `rw` variable storing `*http.ResponseWriter` is renamed to `r` to
  mirror convention in `net/http`