Add HTTP/2 support
Use bradfitz's HTTP/2 library, which is still in development, to enable
HTTP/2 support for user agents that support it.

User agents using HTTP/1.1 can still connect as before.

Note that we only do this for TLS-enabled connections as that's what the
library supports. Chrome and Firefox also do not support HTTP/2 without
TLS.

This is something of an experiment, but given the speed benefits of
HTTP/2 I think it's worth trying.