Add GitHub Oauth
Allow users to log in using GitHub Oauth.

Also adds a `/logout` endpoint.

If a user is not authenticated, we return them to the home page, `/`,
which has a `Log in` button. At present, the `/dashboard` endpoint is
the only one that requires authentication by checking for non-zero user
ID in the dashboard handler (`handler.Dashboard`).