Don't verify SSL certificates when running tests
I'm using a self-signed certificate to run tests, so invoke PhantomJS
with the `--ignore-ssl-errors` option.

Note that the WebDriver specification says:

> WebDriver implementations must support users accessing sites served
> via HTTPS. Access to those sites using self-signed or invalid
> certificates, and where the certificate does not match the serving
> domain must be the same as if the HTTPS was configured properly.

https://w3c.github.io/webdriver/webdriver-spec.html#invalid-ssl-certificates

So we shouldn't have to do this. Selenium has an `acceptSslCerts`
capability we could set but I can't find one for PhantomJS (and indeed,
it should just accept my self-signed certificate).