Check that pages accessed match the expected URL
Sometimes the page accessed does not match the URL requested. For
example, if you access a page secured by a self-signed certificate using
PhantomJS and have not set `--ignore-ssl-errors=true` when invoking
PhantomJS, the page it opens is `about:blank` and no errors are
returned.

By checking explicitly, we can help to avoid this.