Compile version into binary using Makefile
Add a Makefile which sets a version number in the `version` variable in
the `main` package.

If the binary is compiled without the Makefile, the binary will fail
when executed with an error message mentioning the Makefile.

Note that in `main_test.go`, we set a dummy value for the version to
prevent `go test` from failing.