Set Godep GOPATH and PATH in Makefile
Ensure that the GOPATH includes the packages vendored by Godep so that
the vendored versions of package dependencies are used.

This is equivalent to prepending `godep` to the `go test` and `go build`
commands, except that it doesn't require `godep` to be installed just to
build or test the application.