Remove Org ('organisation') model
This model was originally intended to represent organisations, i.e. the
user's employer.

It was added so that some users may represent their organisation as an
administrator, such that they could define things such as the default
number of days of annual leave or leave periods across an organisation,
for example.

Remove it because it's not essential to know who a user works for in
order to record their annual leave and it's just a complication we don't
need just now. This can be revisited when/if a feature requires it.