Add EmployerName field to Job model
Allow users to record the name of the employer, which we can use later
to distinguish between jobs.

Note that we removed the `Org` model in f76c260e for simplicity, though
I think it's okay to add this now as it means we can require new users
to record their employer's name and then later use that to support
multiple jobs per user; also the implementation to record the name is
trivial.