Rename StartDate in LeaveAllowance model
Rename `LeaveAllowance.StartDate` to `LeaveAllowance.StartTime`.

Use the `Time` suffix to denote that fact that these fields are in fact
timestamps.

Also, add an EndTime field to denote when the allowance ends. This is
necessary since as of b4e340f1b, we no longer assume a leave period to
be one calendar year so we must define its duration.