Rename StartDate in LeaveRequest model
Rename `LeaveRequest.StartDate` to `LeaveRequest.StartTime`.

Use the `Time` suffix to denote that fact that these fields are in fact
timestamps.

Also, add an EndTime field to denote when the holiday ends. Having
an end time is necessary because the application has no understanding of
working days, nor locale-specific public holidays. Thus we store the
start and end of the holiday along with the number of working days
(determined by the `Days` field) that actually count as holiday. The
user will need to provide this value.