Add support for PostgreSQL as database
Add support for PostgreSQL, which is configured using the `DB_DIALECT`
and `DB_DATASOURCE` environment variables.

The default remains as an in-memory SQLite3 database, which is useful
for someone wishing to try out the application without having to
configure a database.

I plan to support both SQLite and PostgreSQL for now. If that becomes
too difficult, I'll drop support for SQLite later.

If PostgreSQL is configured using the environment variables, the test
suite will run against PostgreSQL rather than SQLite. I've configured
Travis CI to run tests against both.