Add tests for handler/oauth_callback.go
Add tests for the GitHub Oauth HTTP handler.

These tests stub out a GitHub Oauth response and the GitHub API.

We also check that a new user has been created in the database when a
GitHub user authenticates for the first time. I'm not 100% sure if
checking the database in a unit test is the right thing, but it seems to
be appropriate here. In any case, it's too early to think about
optimising/refactoring this yet.