Flesh out dashboard
Add the necessary logic to:

- let the user enter their employment and leave year details
- see a very basic dashboard which shows past and upcoming leave and how
  many leave days they have remaining this leave year

Uses HTML5 form validation in the user agent where possible. Input
fields for numbers use a regular expression validation pattern to
trigger the numeric keypad on iOS:
http://stackoverflow.com/a/14447832/523704

Current assumptions/limitations present in the handler and model logic:

- user only has one job
- user's job is full-time
- user's leave is calculated annually

...note that the database model itself does not have these same
limitations, so we can improve on this later.

We upgrade Gorm so that we can use its preloading feature:
https://github.com/jinzhu/gorm/tree/25772352c2e1cb6a9c00472999015a32dd646218#preloading-eager-loading