Allow export and sync to calendar
Allow the export of an ICS vCalendar file based on RFC 2445.

The vCalendar file is also accessible via a private URL unique to each
job a user has, such that services such as Google Calendar can
'subscribe' to a user's leave calendar.

The `script-src` and `style-src` modifications in the Content Security
Policy header in `main.go` are necessary to support the modular pop-up
when clicking the 'Add to calendar' link with Javascript enabled.

Specifically, `script-src` uses a hash to match the
`$(document).foundation();` call at the base of the HTML page and
`style-src` allows inline styles, which is currently necessary to
support modernizr.js as used by the Foundation framework:

https://github.com/Modernizr/Modernizr/pull/1263
https://github.com/Modernizr/Modernizr/issues/1450