Settings: Handle updating of existing settings
The settings handler did not previously handle updating existing
settings well.

If the current job ID is non-zero, meaning that the user has already
entered settings previously, we only update the company name and the
date that they started.

Add a `(*Job).Save()` method allowing us to update a job record in the
database.