Add `-dirty` suffix to version if tree is dirty
Add a `-dirty` suffix to the version number that is baked into the
compiled binary if the Git tree is dirty or if the working copy contains
untracked files.

I tried using `git describe --always --dirty`, but it does not handle
the situation where untracked files are present, which is important in
Go because it's possible to add code in new files that would affect the
progam's behaviour.

Adding this suffix should make it easier to diagnose problems
encountered by open source users.

Note that the output of `git status --porcelain` should remain stable
between Git versions according to the official documentation:
http://git-scm.com/docs/git-status