Replace SQLite's total() with something portable
`total()` is specific to SQLite. It differs from `sum()` in that
`total()` will return null if no rows with non-null values are returned.

In our case, this would cause our database queries to fail because `nil`
would be returned instead of the integer we were expecting.

Using `coalesce()`, which is supported by both SQLite and PostgreSQL,
will return the first non-null value it encounters, so zero (0) is
returned if no rows are found.