Dashboard: Fix next leave request date
The "You're next on leave" feature on the dashboard was showing the
leave request furthest in the future, rather than the next from today.

The `(*Job).RequestsLastYearAndFuture()` function returns leave requests
sorted by start date descending, so the next leave request will be the
last item in the `upcomingRequests` slice.