Remove webcal and Google Calendar URLs
Remove the URLs for adding a leave calendar to Google Calendar or using
a webcal URL.

Google Calendar's support for adding calendars via a URL seems flaky at
best:

  - HTTPS URLs are not supported, meaning we'd have to use an insecure
    `http://` `webcal://` URL
  - When testing with a HTTP URL, the calendar did not immediately
    appear, though adding it manually by copying the URL through Google
    Calendar's menus does work

Remove the `webcal` URLs used for Outlook or Apple Calendar because
`webcal` would expose both the calendar secret and the contents of the
calendar in plain-text, meaning anyone sniffing the end-user's traffic
would be able to see when they go on holiday. I think this is a serious
enough privacy concern to warrant removing these. The `webcals` scheme,
which uses TLS, does not appear to be widely supported by browsers at
the current time.

The leave calendar can still be subscribed to using the HTTPS URL for
clients that support it (such as Google Calendar, when adding the URL
manually).