Fix panic if GitHub name is not set
Setting a name in a GitHub account is optional, so not all GitHub
accounts have it set.

Handle this and prevent a nil pointer dereference by using the GitHub
user's username as their name if they have not set their name in GitHub.

We should look at allowing the user to modify their name in LeaveDiary's
setttings page or similar, as the GitHub username may not reflect their
real name at all (using a real name is useful when sharing a leave
calendar etc).