Fix ordering in query to prevent data leak
This query was returning returning results for any job with any end
time, meaning __all__ leave requests belonging to all users where
returned.

Change the order of the `.Where()` and `.Find()` methods to fix. I've
raised an issue in the ORM library:
https://github.com/jinzhu/gorm/issues/495

Before this fix, the SQL query generated was:

    SELECT  * FROM "leave_requests"  WHERE ("leave_requests".deleted_at IS NULL OR "leave_requests".deleted_at <= '0001-01-02') ORDER BY start_time DESC,end_time DESC

After:

    SELECT  * FROM "leave_requests"  WHERE ("leave_requests".deleted_at IS NULL OR "leave_requests".deleted_at <= '0001-01-02') AND ((end_time >= ? AND job_id = ?)) ORDER BY start_time DESC,end_time DESC