Prevent leave request from displaying twice
If a leave request's start time is on the same day as the day a user
views the dashboard, that leave request will show both in 'Past time
off' section and the 'Upcoming' section.

This commit fixes that, so that a leave request starting on today's date
(at 00:00, since we only support full-day leave requests thus far) will
show in the 'Past time off' section.