Add custom recovery middleware for Negroni
Create a new, custom, recovery middleware for Negroni so that:

- we can present a pretty error page to the user if the application
  panics
- we can output the stack trace using Logrus' logger for consistency
  with the rest of the application

Based on Negroni's own recovery middleware:
https://github.com/codegangsta/negroni/blob/c7477ad/recovery.go

I'm a little concerned that my making this recovery code more complex
(by rendering a pretty error page), we're increasing the risk of
panicking while recovering from the original panic.

This might be a good argument for dealing with this higher up the stack
(i.e. by configuring Nginx to output a pretty page when a HTTP 5XX
occurs), but for now I'm not using Nginx so this will do.